% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaTS.R
\name{analyzeNwaTS}
\alias{analyzeNwaTS}
\title{Subnetwork analysis for Time-series data}
\usage{
analyzeNwaTS(nwaList, fdr = 0.001, species, verbose = TRUE)
}
\arguments{
\item{nwaList}{A named list of NWA object}

\item{fdr}{a single numeric value specifying the false discovery for the scoring of nodes
(see BioNet::scoreNodes and Dittrich et al., 2008 for details)}

\item{species}{a single character value specifying the species for which the data should be read.}

\item{verbose}{a single logical value specifying to display detailed messages
(when verbose=TRUE) or not (when verbose=FALSE), default is TRUE.}
}
\value{
In the end, this function will return an updated list of NWA objects.
}
\description{
For each NWA object in 'nwaList', this function will store the subnetwork module identified
by BioNet (if species is given, labels of nodes will also be mapped from
Entrez IDs to gene symbols), and update information about these results to
slot summary of class NWA.
}
\seealso{
\code{\link[HTSanalyzeR2]{analyze}}
}
