% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_preprocess2.R
\name{annotationConvertor2}
\alias{annotationConvertor2}
\title{Convert between different types of gene identifiers}
\usage{
annotationConvertor2(geneList, species = "Dm", initialIDs = "ENTREZID",
  finalIDs = "ENTREZID", keepMultipleMappings = TRUE, verbose = TRUE)
}
\description{
Convert between different types of gene identifiers
}

