% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Biogrid_HS_Mat}
\alias{Biogrid_HS_Mat}
\title{An interactome matrix for \emph{Homo Sapiens}}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 260609 rows and 3 columns.
}
\usage{
Biogrid_HS_Mat
}
\description{
An interactome matrix for \emph{Homo Sapiens}.
}
\examples{
data(Biogrid_HS_Mat)
}
\keyword{datasets}
