% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test.R
\name{cellHTS2OutputStatTests}
\alias{cellHTS2OutputStatTests}
\title{Perform statistical tests on a cellHTS object}
\usage{
cellHTS2OutputStatTests(
  cellHTSobject,
  annotationColumn = "GeneID",
  controls = "neg",
  alternative = "two.sided",
  logged = FALSE,
  tests = "T-test"
)
}
\arguments{
\item{cellHTSobject}{An object of class cellHTS.}

\item{annotationColumn}{A single character value specifying the name of the
column in the fData(cellHTSobject) data frame from which the feature
identifiers will be extracted.}

\item{controls}{A single character value giving the name of the control group
used as the control population in the two-sample tests. If nothing is
specified, the function will look for negative controls labelled "neg".}

\item{alternative}{A single character value specifying the alternative
hypothesis: "two.sided", "less" or "greater".}

\item{logged}{A single logical value specifying whether or not the data has
been logged during the normalization process.}

\item{tests}{A single character value specifying the tests to be performed:
"T-test", "MannWhitney" or "RankProduct". If nothing is specified, all three
tests will be performed. Be aware that the Rank Product test is slower than
the other two, and returns a percent false discovery (equivalent to a FDR,
not a p-value).}
}
\value{
See \code{\link[HTSanalyzeR2]{screenStatTests}}.
}
\description{
Adapter that extracts the measurements, the feature annotation and the
sample/control labelling from a normalized, configured and annotated
`cellHTS` object and passes them to
\code{\link[HTSanalyzeR2]{screenStatTests}}, which performs the tests.
}
\details{
`cellHTS2` has been removed from Bioconductor. This function therefore only
works when that package is installed from a source archive; the tests
themselves are available without it through
\code{\link[HTSanalyzeR2]{screenStatTests}}.
}
\examples{
\dontrun{
data(xn)
test.stats <- cellHTS2OutputStatTests(cellHTSobject = xn,
                                      annotationColumn = "GeneID",
                                      alternative = "two.sided",
                                      tests = c("T-test"))
}
}
