% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_preprocess2.R
\name{duplicateRemover2}
\alias{duplicateRemover2}
\title{This function gets rid of the duplicates in a gene list.}
\usage{
duplicateRemover2(geneList, method = "max")
}
\description{
This function gets rid of the duplicates in a gene list.
}

