% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_enrichmap.R
\docType{methods}
\name{extractEnrichMap,GSCA-method}
\alias{extractEnrichMap,GSCA-method}
\alias{extractEnrichMap}
\title{Extract the enrichment map as an igraph object}
\usage{
\S4method{extractEnrichMap}{GSCA}(object, resultName = "GSEA.results", gscs,
  ntop = NULL, allSig = TRUE, gsNameType = "id", specificGeneset = NULL,
  cutoff = NULL)
}
\arguments{
\item{object}{A GSCA object.}

\item{resultName}{A single character value specifying draw an enrichment map based on
which result. Could only be 'HyperGeo.results' or 'GSEA.results'.}

\item{gscs}{A character vector specifying the names of gene set collections
of which the top significant gene sets will be plotted.}

\item{ntop}{A single integer or numeric value specifying how many gene
sets of top significance will be plotted.}

\item{allSig}{A single logical value. If 'TRUE', all significant gene sets
(GSEA adjusted p-value < 'pValueCutoff' of slot 'para') will be used; otherwise,
only top 'ntop' gene sets will be used.}

\item{gsNameType}{A single character value specifying the type of the gene set names that
will be displayed as the names of nodes in the enrichment map. The type of the gene
set names should be one of the following: "id", "term" or "none".}

\item{specificGeneset}{A named list of specific gene sets. Specifically, this term needs to be
a subset of all analyzed gene sets which can be roughly gotten by
\strong{getTopGeneSets(object, resultName, gscs, ntop = 20000, allSig = FALSE)}.}

\item{cutoff}{A numeric value between 0 and 1. This parameter is setted as a cutoff of edge weight in the enrichment
map for better visualization. When the edge weight, namely the Jaccard coefficient between two gene sets, is less than
this cutoff, this edge would not be showed in the enrichment map.
The order of the list must match the order of results gotten by aboved function \strong{getTopGeneSets}.}
}
\value{
An object of igraph with all attributes about the enrichement map.
}
\description{
Extract the enrichment map from an analyzed GSCA object as igraph object for further external using.
Users can also modify the igraph object.
}
\examples{
## load a GSCA object(see the examples of 'analyze' GSCA for details)
library(igraph)
data(d7_gsca)

## extract the enrichment map for top 30 significant 'GO_MF' gene sets of GSEA results in an igraph object
extractEnrichMap(d7_gsca, resultName = "GSEA.results", gscs=c("GO_MF"),
                allSig = FALSE, ntop = 30, gsNameType = "term")

}
