% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_enrichmap.R
\name{fetchGSCASeriesValues}
\alias{fetchGSCASeriesValues}
\title{FetchGSCASeriesValues}
\usage{
fetchGSCASeriesValues(gscaObjs, resultName = "GSEA.results", gscs,
  ntop = NULL, allSig = TRUE, gsNameType = "id")
}
\description{
FetchGSCASeriesValues
}
