% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwa_view.R
\name{fetchNWASeriesValues}
\alias{fetchNWASeriesValues}
\title{Extract and combine subnetwork information of given NWA objects.}
\usage{
fetchNWASeriesValues(nwaObjs)
}
\description{
This method is internally used for force-graph drawing. The subnetwork data of all the NWA objects
are extacted and combined into corresponding dataframes. The attributes that are changes with the time
are also returned.
}
