% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_widget.R
\name{forceGraph}
\alias{forceGraph}
\title{Create an interactive force-directed network widget}
\usage{
forceGraph(
  nodes,
  links,
  nMappings,
  lMappings,
  options,
  width = NULL,
  height = NULL,
  seriesData = NULL
)
}
\arguments{
\item{nodes}{the node information of the force-directed graph}

\item{links}{the link information of the force-directed graph}

\item{nMappings}{indicates which attributes are used for node rendering. The
available render options are "id", "size", "color", "scheme", "label",
"label_id" and "label_term".}

\item{lMappings}{indicates which attributes are used for link rendering. The
available render options are "source", "target" and "weight".}

\item{options}{rendering options such as the title, legend title, the label
type ("id", "term" or "none") and the colour scaler ("log10" or "linear").}

\item{width}{the width of the widget}

\item{height}{the height of the widget}

\item{seriesData}{a character vector of time points for time-series data}
}
\value{
An object of class 'htmlwidget' (a 'visNetwork' graph).
}
\description{
Builds an interactive network graph for enrichment maps and enriched
subnetworks. Rendering is done by \pkg{visNetwork} (vis.js); the widget works
in R Markdown documents, in the Shiny report produced by
\code{\link[HTSanalyzeR2]{report}} and as a standalone HTML page.
}
