% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{getInteractome,NWA-method}
\alias{getInteractome,NWA-method}
\alias{getInteractome}
\title{Accessors for the 'interactome' slot of a 'NWA' object.}
\usage{
\S4method{getInteractome}{NWA}(object)
}
\arguments{
\item{object}{An object of 'NWA'.}
}
\value{
This function will return interactome used in 'NWA' analysis.
}
\description{
This function get the interactome used in 'NWA' analysis.
}
\examples{
data(d7_nwa)
s1 <- getInteractome(d7_nwa)
}
