% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{getPara,GSCA-method}
\alias{getPara,GSCA-method}
\alias{getPara}
\alias{getPara,NWA-method}
\title{Accessors for the 'para' slot of a 'GSCA' or 'fdr' slot of a 'NWA' object.}
\usage{
\S4method{getPara}{GSCA}(object)

\S4method{getPara}{NWA}(object)
}
\arguments{
\item{object}{An object of 'GSCA' or 'NWA'.}
}
\value{
This function will return all the parameters.
}
\description{
This function get all the parameters used in 'GSCA' or 'NWA' analysis.
}
\examples{
# ===========================================================
# GSCA class
data(d7_gsca)
para1 <- getPara(d7_gsca)

# ===========================================================
# NWA class
data(d7_nwa)
para1 <- getPara(d7_nwa)

}
