% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\docType{methods}
\name{getSummary,GSCA-method}
\alias{getSummary,GSCA-method}
\alias{getSummary}
\alias{getSummary,NWA-method}
\title{Accessors for the 'summary' slot of a 'GSCA' or 'NWA' object.}
\usage{
\S4method{getSummary}{GSCA}(object)

\S4method{getSummary}{NWA}(object)
}
\arguments{
\item{object}{An object of 'GSCA' or 'NWA'.}
}
\value{
This function will return the summary information.
}
\description{
This 'summary' slot summarized an analyzed 'GSCA' or 'NWA' object.
}
\examples{
# ===========================================================
# GSCA class
data(d7_gsca)
s1 <- getSummary(d7_gsca)
# ===========================================================
# NWA class
data(d7_nwa)
s1 <- getSummary(d7_nwa)

}
