% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_class.R
\name{GSCA}
\alias{GSCA}
\title{An S4 class for Gene Set Collection Analyses on high-throughput screens}
\usage{
GSCA(listOfGeneSetCollections, geneList, hits)
}
\description{
An S4 class for Gene Set Collection Analyses on high-throughput screens
}
\section{Slots}{

\describe{
\item{\code{listOfGeneSetCollections}}{A list of gene set collections (a 'gene
set collection' is a list of gene sets).}

\item{\code{geneList}}{A numeric or integer vector of phenotypes named by gene
identifiers.}

\item{\code{hits}}{A character vector of the gene identifiers (used as hits in
the hypergeometric tests).}
}}

