% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_class.R
\docType{class}
\name{GSCABatch-class}
\alias{GSCABatch-class}
\title{An S4 class for Time series data package in Gene Set Collection Analyses on high-throughput screens}
\usage{
new("GSCABatch", expInfor, listOfGeneSetCollections, phenotypeTS, hitsTS = list())
}
\description{
This S4 class packages time-series data for further GSCA. To put it more clearly, it'll finally generate a list of
GSCA objects for further analyses.
}
\section{Slots}{

\describe{
\item{\code{expInfor}}{A character matrix contains experiment information with each experiment in row and information in column.
Should at least contain two columns named as 'ID' and 'Desription'.}

\item{\code{listOfGeneSetCollections}}{A list of gene set collections (a 'gene
set collection' is a list of gene sets).}

\item{\code{phenotypeTS}}{A list of phenotypes, each element of this list is a numeric vector phenotypes named by gene
identifiers for each time point. Note: the order of each element of this list must match the order of 'expInfor' ID.}

\item{\code{hitsTS}}{A list of hits, each element is a character vector of the gene identifiers (used as hits in
the hypergeometric tests). It's needed if you want do GSOA. Note: the order of each element of this list
 must match the order of 'expInfor' ID.}

\item{\code{listOfGSCA}}{A list of initialized GSCA object for futher GSCA.}
}}

\examples{
data(d7, d13, d25)

## generate expInfor to describe the information of time series data
expInfor <- matrix(c("d7", "d13", "d25"), nrow = 3, ncol = 2,
                   byrow = F, dimnames = list(NULL, c("ID", "Description")))

## package phenotypeTS into a list of phenotypes
datalist <- list(d7, d13, d25)
phenotypeTS <- lapply(datalist, function(x) {
                      tmp <- as.vector(x$neg.lfc)
                      names(tmp) <- x$id
                      tmp})

## set up a list of gene set collections
GO_MF <- GOGeneSets(species="Hs", ontologies=c("MF"))
ListGSC <- list(GO_MF=GO_MF)

## package hitsTS if you also want to do GSOA, otherwise ignore it
hitsTS <- lapply(datalist, function(x){
tmp <- x[x$neg.p.value < 0.01, "id"]
tmp})

## Example1: create an object of class GSCABatch with hitsTS
GSCABatch <- new("GSCABatch", expInfor = expInfor, phenotypeTS = phenotypeTS,
                 listOfGeneSetCollections = ListGSC, hitsTS = hitsTS)
GSCABatch

## Example2: create an object of class GSCABatch without hitsTS
GSCABatch <- new("GSCABatch", expInfor = expInfor, phenotypeTS = phenotypeTS,
                 listOfGeneSetCollections = ListGSC)
GSCABatch
}
