% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_class.R
\name{GSCABatch}
\alias{GSCABatch}
\title{An S4 class for Time series data package in Gene Set Collection Analyses on high-throughput screens}
\usage{
GSCABatch(expInfor, listOfGeneSetCollections, phenotypeTS, hitsTS = list())
}
\description{
This S4 class packages time-series data for further GSCA. To put it more clearly, it'll finally generate a list of
GSCA objects for further analyses.
}
\section{Slots}{

\describe{
\item{\code{expInfor}}{A character matrix contains experiment information with each experiment in row and information in column.
Should at least contain two columns named as 'ID' and 'Desription'.}

\item{\code{listOfGeneSetCollections}}{A list of gene set collections (a 'gene
set collection' is a list of gene sets).}

\item{\code{phenotypeTS}}{A list of phenotypes, each element of this list is a numeric vector phenotypes named by gene
identifiers for each time point. Note: the order of each element of this list must match the order of 'expInfor' ID.}

\item{\code{hitsTS}}{A list of hits, each element is a character vector of the gene identifiers (used as hits in
the hypergeometric tests). It's needed if you want do GSOA. Note: the order of each element of this list
 must match the order of 'expInfor' ID.}

\item{\code{listOfGSCA}}{A list of initialized GSCA object for futher GSCA.}
}}

