% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gscaTS}
\alias{gscaTS}
\title{A list of 'GSCA' object}
\format{An object of class \code{list} of length 3.}
\usage{
gscaTS
}
\description{
A list of 'GSCA' object generated by time-series analysis using \emph{HTSanalyzeR2}.
}
\keyword{datasets}
