% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE33113_exp}
\alias{GSE33113_exp}
\title{Expression of a demo data: GSE33113}
\format{An object of class \code{data.frame} with 21656 rows and 90 columns.}
\usage{
GSE33113_exp
}
\description{
Expression of normalized data GSE33113 in vignette, which contains 21656 genes
expression among 78 colon rectal cancer patients.
}
\keyword{datasets}
