% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE33113_label}
\alias{GSE33113_label}
\title{Concensus molecular subtype label of GSE33113}
\format{An object of class \code{character} of length 90.}
\usage{
GSE33113_label
}
\description{
Concensus molecular subtype label of GSE33113
}
\keyword{datasets}
