% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE33113_limma}
\alias{GSE33113_limma}
\title{limma result of a gene expression data}
\format{An object of class \code{data.frame} with 21656 rows and 6 columns.}
\usage{
GSE33113_limma
}
\description{
The differential expressed analysis result between CMS4 colon cancer subtype and
non-CMS4 of a gene expression dataset with GEO number-"GSE33113" using R package \emph{limma}.
}
\examples{
data(GSE33113_limma)

}
\keyword{datasets}
