% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_load_genesets.R
\name{KeggGeneSets}
\alias{KeggGeneSets}
\title{Create a list of KEGG gene sets}
\usage{
KeggGeneSets(species = "Hs")
}
\arguments{
\item{species}{A single character value specifying a choice of species,
such as "Dm" ("Drosophila_melanogaster"), "Hs" ("Homo_sapiens"),
"Rn" ("Rattus_norvegicus") or "Mm" ("Mus_musculus").}
}
\value{
A list of gene sets, with names as KEGG pathway IDs. Each gene
set is a group of genes represented by Entrez identifiers.
}
\description{
This function creates a list of gene sets based on KEGG pathways terms.
It is species-specific, and returns a list of gene sets, each of which
is a character vector of Entrez gene identifiers.
}
\details{
This function needs Internet connection and relies on the
following packages: KEGGREST.
}
\examples{
\dontrun{
library(KEGGREST)
library(org.Hs.eg.db)
HS_KEGG <- KeggGeneSets(species = "Hs")
}
}
\seealso{
\code{\link[HTSanalyzeR2]{GOGeneSets}}, \code{\link[HTSanalyzeR2]{MSigDBGeneSets}}
}
