% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_enrichmap.R
\name{makeEmptyGraph}
\alias{makeEmptyGraph}
\title{Generate an empty igraph object with given vertex/edge attributes}
\usage{
makeEmptyGraph(NAttributes, EAttributes)
}
\description{
Generate an empty igraph object with given vertex/edge attributes
}
