% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_load_genesets.R
\name{MSigDBGeneSets}
\alias{MSigDBGeneSets}
\title{Create a list of gene sets for specific species from MSigDB database}
\usage{
MSigDBGeneSets(species = "Hs", collection = "C2", subcategory = NULL)
}
\arguments{
\item{species}{A single character value specifying the species of the gene sets of MSigDB.
Now we support 10 species: 'Bt'(Bos taurus), 'Ce'(Caenorhabditis elegans),
'Cfa'(Canis lupus familiaris), 'Dm'(Drosophila melanogaster), 'Dr'(Danio rerio),
'Gg'(Gallus gallus), 'Hs'(Homo sapiens), 'Mm'(Mus musculus), 'Rn'(Rattus norvegicus),
'Sc'(Saccharomyces cerevisiae) and 'Ss'(Sus scrofa).}

\item{collection}{A single character value specifying a choice of collection.
Valid values include 'H'(hallmark gene sets), 'C1'(positional gene sets),
'C2'(curated gene sets), 'C3'(motif gene sets), 'C4'(computational gene sets),
'C5'(GO gene sets), 'C6'(oncogenic signatures), 'C7'(immunologic signatures).
More details please refer to
\href{https://software.broadinstitute.org/gsea/msigdb}{MSigDB}.}

\item{subcategory}{A single character value or NULL specifying a subcategory of
the selected MSigDB collection. See \code{\link[msigdbr]{msigdbr}} for details.}
}
\value{
Return a list of gene sets of specific collection in
\href{https://software.broadinstitute.org/gsea/msigdb}{MSigDB} of version 6.2.
}
\description{
This function creates gene set collections based on MSigDB database version 6.2. Currently
our package supports all 8 collections for 10 species retrieved by \code{\link[msigdbr]{msigdbr}}.
It returns a list of gene sets collections
with the elements of the gene sets represented by Entrez Gene IDs.
}
\examples{
C2_MSig <- MSigDBGeneSets(species = "Hs", collection = "C2", subcategory = NULL)
}
\seealso{
\code{\link[HTSanalyzeR2]{GOGeneSets}}, \code{\link[HTSanalyzeR2]{KeggGeneSets}}
}
