% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwa_analyze.R
\name{networkAnalysis}
\alias{networkAnalysis}
\title{Identify enriched subnetworks}
\usage{
networkAnalysis(pvalues, graph, fdr = 0.001, verbose = TRUE)
}
\arguments{
\item{pvalues}{a numeric vector of p-values}

\item{graph}{an object of class igraph, used as the interactome in
the network analysis}

\item{fdr}{a single numeric value specifying the false discovery for
the scoring of nodes (see BioNet::scoreNodes and Dittrich et al., 2008
for details)}

\item{verbose}{a single logical value indicating to display detailed
messages (when verbose=TRUE) or not (when verbose=FALSE)}
}
\value{
a subnetwork module of class igraph
}
\description{
This function finds subnetworks enriched for genes with significant
phenotypes based on the package 'BioNet'.
}
\details{
This function takes in a vector of p-values and a graph standing
for the interactome to identify the maximum scoring subnetwork based on
the BioNet package.
}
