% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwa_view.R
\name{networkPlot}
\alias{networkPlot}
\title{filepath' with the name 'filename'.
diff.expr' holds the phenotype for the nodes of the sub-network
present' contains the information of wether a node has an}
\usage{
networkPlot(nwAnalysisOutput, phenotypeVector = NULL)
}
\description{
filepath' with the name 'filename'.
diff.expr' holds the phenotype for the nodes of the sub-network
present' contains the information of wether a node has an
}

