% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwa}
\alias{nwa}
\title{An object of class 'NWA'}
\format{An object of class \code{NWA} of length 1.}
\usage{
nwa
}
\description{
An object of class 'NWA' generated by \emph{HTSanalyzeR2}.
}
\keyword{datasets}
