% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_class.R
\name{NWABatch}
\alias{NWABatch}
\title{An S4 class for Time series data package in NetWork Analysis on high-throughput screens}
\usage{
NWABatch(expInfor, pvalueTS, phenotypeTS = list(), interactome = NA)
}
\description{
This S4 class packages time-series data for further NWA.
}
\section{Slots}{

\describe{
\item{\code{expInfor}}{A character matrix contains experiment information with each experiment in row and information in column.
At least contains two columns named as 'ID' and 'Desription'.}

\item{\code{phenotypeTS}}{A list of phenotypes, each element of this list is a numeric vector phenotypes named by gene
identifiers for each time point. Note: the order of each element of this list must match the order of 'expInfor' ID.}

\item{\code{pvalueTS}}{A list of pvalues, each element of this list is a numeric vector pvalues named by gene
identifiers for each time point. Note: the order of each element of this list must match the order of 'expInfor' ID.}

\item{\code{interactome}}{object of class igraph.}
}}

