% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaTS.R
\name{nwaTS}
\alias{nwaTS}
\title{Produce a list of NWA object for NetWork Analysis on high-throughput screens}
\usage{
nwaTS(object, interactome = NA)
}
\arguments{
\item{object}{A TSImport object.}

\item{interactome}{an object of class igraph.}
}
\value{
In the end, this function will return a named list of new NWA object.
}
\description{
This function will produce a list of NWA object for time-series data to do network analysis
}
\seealso{
\code{\link[HTSanalyzeR2]{NWA}}
}
