% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_view.R
\docType{methods}
\name{plotGSEA,GSCA-method}
\alias{plotGSEA,GSCA-method}
\title{Plot and save figures of GSEA results for top significant gene sets}
\usage{
\S4method{plotGSEA}{GSCA}(object, gscs, ntop = NULL, allSig = FALSE,
  filepath = ".", output = "pdf", ...)
}
\arguments{
\item{object}{A GSCA object.}

\item{gscs}{A character vector specifying the names of gene set collections whose top
significant gene sets will be plotted.}

\item{ntop}{A single integer or numeric value specifying how many gene sets of top
significance will be plotted.}

\item{allSig}{A single logical value. If 'TRUE', all significant gene sets (GSEA adjusted
p-value < 'pValueCutoff' of slot 'para') will be plotted; otherwise, only top 'ntop' gene
 sets will be plotted.}

\item{filepath}{A single character value specifying where to store GSEA figures.}

\item{output}{A single character value specifying the format of output image: "pdf" or "png".}

\item{...}{Other arguments used by the function png or pdf such as 'width' and 'height'}
}
\description{
This is a generic function.When implemented as the S4 method for objects of class GSCA,
this function plots figures of the positions of genes of the gene set in the ranked gene
list and the location of the enrichment score for top significant gene sets.
}
\examples{
library(org.Hs.eg.db)
library(GO.db)
library(KEGGREST)
## load data for enrichment analyses
data(d7)
phenotype <- as.vector(d7$neg.lfc)
names(phenotype) <- d7$id

## select hits if you also want to do GSOA, otherwise ignore it
hits <-  names(phenotype[which(abs(phenotype) > 2)])

## set up a list of gene set collections
GO_MF <- GOGeneSets(species="Hs", ontologies=c("MF"))
PW_KEGG <- KeggGeneSets(species="Hs")
ListGSC <- list(GO_MF=GO_MF, PW_KEGG=PW_KEGG)

## create an object of class 'GSCA'
gsca <- new("GSCA", listOfGeneSetCollections = ListGSC, geneList = phenotype, hits = hits)

## do preprocessing
gsca <- preprocess(gsca, species="Hs", initialIDs="SYMBOL", keepMultipleMappings=TRUE,
                   duplicateRemoverMethod="max", orderAbsValue=FALSE)

## support parallel calculation using doParallel package
doParallel::registerDoParallel(cores=4)

## do hypergeometric tests and GSEA
gsca <- analyze(gsca, para=list(pValueCutoff=0.05, pAdjustMethod ="BH",
                                nPermutations=100, minGeneSetSize=200, exponent=1),
                                doGSOA = TRUE, doGSEA = TRUE)

## summarize gsca
summarize(gsca)

## plot  significant gene sets in GO_MF and PW_KEGG
plotGSEA(gsca, gscs=c("GO_MF","PW_KEGG"), ntop=3, filepath=".")
}
