% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscaTS.R
\name{preprocessGscaTS}
\alias{preprocessGscaTS}
\title{A preprocessing method for a GSCABatch object of Time-series data Gene Set Collection Analyses}
\usage{
preprocessGscaTS(object, species = "Hs", initialIDs = "SYMBOL",
  keepMultipleMappings = TRUE, duplicateRemoverMethod = "max",
  orderAbsValue = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{A GSCABatch object.}

\item{species}{A single character value specifying the species for which the
data should be read.}

\item{initialIDs}{A single character value specifying the type of initial
identifiers for input gscaList}

\item{keepMultipleMappings}{A single logical value. If TRUE, the function
keeps the entries with multiple mappings (first mapping is kept). If FALSE,
the entries with multiple mappings will be discarded.}

\item{duplicateRemoverMethod}{A single character value specifying the method
to remove the duplicates. See duplicateRemover for details.}

\item{orderAbsValue}{A single logical value indicating whether the values
should be converted to absolute values and then ordered (if TRUE), or
ordered as they are (if FALSE).}

\item{verbose}{A single logical value specifying to display detailed messages
(when verbose=TRUE) or not (when verbose=FALSE), default is TRUE.}
}
\value{
In the end, this function will return an updated list of GSCA object.
}
\description{
This function will do basic preprocessing for each GSCA object of 'GSCABatch'.
}
\seealso{
\code{\link[HTSanalyzeR2]{preprocess}}
}
