% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaTS.R
\name{preprocessNwaTS}
\alias{preprocessNwaTS}
\title{A preprocessing method for a list of NWA object of Time-series data}
\usage{
preprocessNwaTS(nwaList, species = "Hs", initialIDs = "SYMBOL",
  keepMultipleMappings = TRUE, duplicateRemoverMethod = "max",
  verbose = TRUE)
}
\arguments{
\item{nwaList}{A named list of NWA object}

\item{species}{A single character value specifying the species for which the
data should be read.}

\item{initialIDs}{A single character value specifying the type of initial
identifiers for input nwaList.}

\item{keepMultipleMappings}{A single logical value. If TRUE, the function
keeps the entries with multiple mappings (first mapping is kept). If FALSE,
the entries with multiple mappings will be discarded.}

\item{duplicateRemoverMethod}{A single character value specifying the method
to remove the duplicates. See duplicateRemover for details.}

\item{verbose}{A single logical value specifying to display detailed messages
(when verbose=TRUE) or not (when verbose=FALSE), default is TRUE.}
}
\value{
In the end, this function will return an updated list of NWA object.
}
\description{
This function will do basic preprocessing for each NWA object of 'nwaList'.
}
\seealso{
\code{\link[HTSanalyzeR2]{preprocess}}
}
