% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_report.R
\docType{methods}
\name{report,GSCA-method}
\alias{report,GSCA-method}
\title{Write HTML reports for enrichment or network analyses}
\usage{
\S4method{report}{GSCA}(object, nodeOptions = list(),
  reportDir = "GSCAReport")
}
\arguments{
\item{object}{an object. When this function is implemented as the S4
method of class 'GSCA' or 'NWA', this argument is an object of class
'GSCA' or 'NWA'.}

\item{nodeOptions}{a list of interested gene set terms which can be edited in the shiny report such as
changing the shape of this set, etc.}

\item{reportDir}{a single character value specifying the directory to store reports. For default
the enrichment analysis reports will be stored in the directory called "GSCAReport"}
}
\description{
This is a generic function.
When implemented as the method of class GSCA or NWA, this function produces reports for
either the Gene Set Collection Analysis or the Network Analysis.
}
\examples{
## Not run:
library(org.Dm.eg.db)
library(GO.db)
## load data for enrichment analyses
data(data4enrich)
## select hits
hits <- names(data4enrich)[abs(data4enrich) > 2]
## set up a list of gene set collections
GO_MF <- GOGeneSets(species="Dm", ontologies=c("MF"))
ListGSC <- list(GO_MF=GO_MF)
## create an object of class 'GSCA'
gsca <- GSCA(listOfGeneSetCollections = ListGSC, geneList = data4enrich, hits = hits)
## print gsca
gsca
## do preprocessing
gsca <- preprocess(gsca, species="Dm", initialIDs="FLYBASECG", keepMultipleMappings=TRUE, duplicateRemoverMethod="max", orderAbsValue=FALSE)
## do hypergeometric tests and GSEA
gsca <- analyze(gsca, para=list(pValueCutoff=0.05, pAdjustMethod ="BH", nPermutations=100, minGeneSetSize=200, exponent=1))
summarize(gsca)
## append gene set terms to results
gsca <- appendGSTerms(gsca, goGSCs=c("GO_MF"), keggGSCs= NULL, msigdbGSCs=NULL))
## report for a GSCA object
report(gsca)




}

