% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_report.R
\name{reportAll}
\alias{reportAll}
\title{Write HTML reports for both enrichment and network analyses}
\usage{
reportAll(gsca, nwa, gscaNodeOptions = NULL, nwaNodeOptions = NULL,
  reportDir = "AnalysisReport")
}
\arguments{
\item{gsca}{an object of class GSCA}

\item{nwa}{an object of class NWA}

\item{gscaNodeOptions}{a list of interested gene set terms which can be edited in the shiny report such as
changing the shape of this set, etc.}

\item{nwaNodeOptions}{a list of interested phenotypes which can be edited in the shiny report such as
changing the shape of this set, etc.}

\item{reportDir}{a single character value specifying the directory to store reports. For default both the
enrichment analysis and network analysis reports will be stored in the directory called "AnalysisReport"}
}
\description{
This function can create HTML reports for both gene sets enrichment analysis and network analysis using
shiny.
}
