% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwa_view.R
\docType{methods}
\name{viewSubNet,NWA-method}
\alias{viewSubNet,NWA-method}
\alias{viewSubNet}
\title{Plot the identified subnetwork of an NWA object}
\usage{
\S4method{viewSubNet}{NWA}(object, options = list(), seriesObjs = NULL)
}
\arguments{
\item{object}{An NWA object.}

\item{options}{A list of options to modify the enrichmentmap. Details are not showed
here due to too many options. Users are highly recommended to modify the enrichment
map in a shiny report by  \code{\link[HTSanalyzeR2]{report}}.}

\item{seriesObjs}{A list of NWA object. Internally used in the shiny report for visualizing
the subnetwork of time series data. No need to explicitly set it!}
}
\value{
In the end, this function would plot the identified subnetwork.
}
\description{
Plot the identified subnetwork of an NWA object.
}
\examples{
\dontrun{
## load a NWA object(see the examples of analyze NWA for details)
data(d7_nwa)

## plot the subnetwork
viewSubNet(d7_nwa)
}
}
