% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xn}
\alias{xn}
\title{An object of class 'cellHTS2'}
\format{
An object of class \code{cellHTS} with 21888 rows and 2 columns.
}
\usage{
xn
}
\description{
An object of class 'cellHTS2' which has been normalized, configured and annotated.
}
\examples{
\dontrun{
data(xn)
}
}
\keyword{datasets}
