% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_analyze.R, R/nwa_analyze.R
\docType{methods}
\name{analyze,GSCA-method}
\alias{analyze,GSCA-method}
\alias{analyze,NWA-method}
\title{Gene Set Collection Analysis or NetWork Analysis}
\usage{
\S4method{analyze}{GSCA}(object, para = list(pValueCutoff = 0.05,
  pAdjustMethod = "BH", nPermutations = 1000, minGeneSetSize = 15, exponent =
  1), verbose = TRUE, doGSOA = FALSE, doGSEA = TRUE)

\S4method{analyze}{NWA}(object, fdr = 0.001, species, verbose = TRUE)
}
\arguments{
\item{object}{An object. When this function is implemented as the S4
method of class 'GSCA' or 'NWA', this argument is an object of class
'GSCA' or 'NWA'.}

\item{para}{A list of parameters for GSEA and hypergeometric tests.
This argument is only for class GSCA.}

\item{verbose}{A single logical value specifying to display detailed messages
(when verbose=TRUE) or not (when verbose=FALSE).}

\item{doGSOA}{A single logical value specifying to perform gene set
overrepresentation analysis(hypergeometric test) (when doGSOA=TRUE) or not (when doGSOA=FALSE).}

\item{doGSEA}{A single logical value specifying to perform gene set
enrichment analysis (when doGSEA=TRUE) or not (when doGSEA=FALSE).}

\item{fdr}{a single numeric value specifying the false discovery for the scoring of nodes
(see BioNet::scoreNodes and Dittrich et al., 2008 for details)}

\item{species}{a single character value specifying the species for which the data should be read.}

\item{para$pValueCutoff}{A single numeric value specifying the cutoff for adjusted p-values considered
significant.}

\item{para$pAdjustMethod}{A single character value specifying the p-value adjustment method to be used
(see 'p.adjust' for details).}

\item{para$nPermutations}{A single integer or numeric value specifying the number of permutations for
deriving p-values in GSEA.}

\item{para$minGeneSetSize}{A single integer or numeric value specifying the minimum number of elements
shared by a gene set and the input total genes. Gene sets with
fewer than this number are removed from both hypergeometric analysis and GSEA.}

\item{para$exponent}{A single integer or numeric value used in weighting phenotypes in GSEA.}
}
\value{
In the end, this function will return an updated object of
class GSCA or NWA. All the analyzed results could be found in slot \emph{result}.
}
\description{
This is a generic function.
}
\section{Methods (by class)}{
\itemize{
\item \code{GSCA}: The function will perform gene set collection analysis
on a GSCA object and update information about these results to slot
\emph{summary} of class GSCA.

\item \code{NWA}: The function will store the subnetwork module identified
by BioNet (if species is given, labels of nodes will also be mapped from
Entrez IDs to gene symbols), and update information about these results to
slot summary of class NWA.
}}

\examples{
# ====================================================
# Gene Set Collection Analysis Part
library(org.Hs.eg.db)
library(GO.db)
library(KEGGREST)
## load data for enrichment analyses
data(d7)
phenotype <- as.vector(d7$neg.lfc)
names(phenotype) <- d7$id

## select hits if you also want to do GSOA, otherwise ignore it
hits <-  names(phenotype[which(abs(phenotype) > 2)])

## set up a list of gene set collections
GO_MF <- GOGeneSets(species="Hs", ontologies=c("MF"))
PW_KEGG <- KeggGeneSets(species="Hs")
ListGSC <- list(GO_MF=GO_MF, PW_KEGG=PW_KEGG)

## create an object of class 'GSCA'
gsca <- new("GSCA", listOfGeneSetCollections = ListGSC, geneList = phenotype, hits = hits)

## do preprocessing
gsca <- preprocess(gsca, species="Hs", initialIDs="SYMBOL", keepMultipleMappings=TRUE,
                   duplicateRemoverMethod="max", orderAbsValue=FALSE)

## support parallel calculation using doParallel package
doParallel::registerDoParallel(cores=4)

## do hypergeometric tests and GSEA
gsca <- analyze(gsca, para=list(pValueCutoff=0.05, pAdjustMethod ="BH",
                                nPermutations=100, minGeneSetSize=200, exponent=1),
                                doGSOA = TRUE, doGSEA = TRUE)

## summarize gsca
summarize(gsca)
head(gsca@result$GSEA.results$GO_MF)
head(gsca@result$HyperGeo.results$PW_KEGG)
# Conducting the preparation steps
data(xn)
data(data4enrich)
# Conducting one sample t-test & compute the p-values
test.stats <- cellHTS2OutputStatTests(cellHTSobject=xn,annotationColumn="GeneID", alternative="two.sided",tests=c("T-test"))
library(BioNet)
pvalues <- BioNet::aggrPvals(test.stats, order=2, plot=FALSE)
nwa <- NWA(pvalues=pvalues, phenotypes=data4enrich)
nwa <- preprocess(nwa, species="Dm", initialIDs="FLYBASECG", keepMultipleMappings=TRUE, duplicateRemoverMethod="max")
nwa_inter <- interactome(nwa, species="Dm", reportDir="biogrid", genetic=FALSE)
# Case1: Conducting the network analysis step
nwa_result <- analyze(nwa_inter, fdr=0.0001, species="Dm")
# Case2: Conducting the network analysis step(with the interactionMatrix)
data(nwam)
nwam_inter <- interactome(nwam, species="Dm", reportDir="biogrid", genetic=FALSE)
nwam_result <- analyze(nwam_inter, fdr=0.0001, species="Dm")
}
