% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_analyze.R, R/nwa_analyze.R
\docType{methods}
\name{analyze,GSCA-method}
\alias{analyze,GSCA-method}
\alias{analyze,NWA-method}
\title{Gene Set Collection Analysis or NetWork Analysis}
\usage{
\S4method{analyze}{GSCA}(object, para = list(pValueCutoff = 0.05,
  pAdjustMethod = "BH", nPermutations = 1000, minGeneSetSize = 15, exponent =
  1), verbose = TRUE, doGSOA = FALSE, doGSEA = TRUE)

\S4method{analyze}{NWA}(object, fdr = 0.001, species, verbose = TRUE)
}
\arguments{
\item{object}{an object. When this function is implemented as the S4
method of class 'GSCA' or 'NWA', this argument is an object of class
'GSCA' or 'NWA'.}

\item{para}{a list of parameters for GSEA and hypergeometric tests.}

\item{verbose}{a single logical value specifying to display detailed messages
(when verbose=TRUE) or not (when verbose=FALSE)}

\item{doGSOA}{a single logical value specifying to perform gene set
overrepresentation analysis (when doGSOA=TRUE) or not (when doGSOA=FALSE)}

\item{doGSEA}{a single logical value specifying to perform gene set
enrichment analysis (when doGSEA=TRUE) or not (when doGSEA=FALSE)}

\item{fdr}{a single numeric value specifying the false discovery for the scoring of nodes
(see BioNet::scoreNodes and Dittrich et al., 2008 for details)}

\item{species}{a single character value specifying the species for which the data should be read.}

\item{para$pValueCutoff}{a single numeric value specifying the cutoff for p-values considered
significant}

\item{para$pAdjustMethod}{a single character value specifying the p-value adjustment method to be used
(see 'p.adjust' for details)}

\item{para$nPermutations}{a single integer or numeric value specifying the number of permutations for
deriving p-values in GSEA}

\item{para$minGeneSetSize}{a single integer or numeric value specifying the minimum number of elements
in a gene set that must map to elements of the gene universe. Gene sets with
fewer than this number are removed from both hypergeometric analysis and GSEA.}

\item{para$exponent}{a single integer or numeric value used in weighting phenotypes in GSEA.}
}
\value{
In the end, this function will return an updated object of
class GSCA or NWA.
}
\description{
This is a generic function.
}
\section{Methods (by class)}{
\itemize{
\item \code{GSCA}: The function will store the results from function
analyzeGeneSetCollections in slot result, and update information about
these results to slot summary of class GSCA.

\item \code{NWA}: The function will store the subnetwork module identified
by BioNet (if species is given, labels of nodes will also be mapped from
Entrez IDs to gene symbols), and update information about these results to
slot summary of class NWA.
}}

\examples{
# ====================================================
# Gene Set Collection Analysis Part
## Not run:
library(org.Dm.eg.db)
library(GO.db)
## load data for enrichment analyses
data(data4enrich)
## select hits
hits <- names(data4enrich)[abs(data4enrich) > 2]
## set up a list of gene set collections
GO_MF <- GOGeneSets(species="Dm", ontologies=c("MF"))
ListGSC <- list(GO_MF=GO_MF)
## create an object of class 'GSCA'
gsca <- GSCA(listOfGeneSetCollections = ListGSC, geneList = data4enrich, hits = hits)
## print gsca
gsca
## do preprocessing
gsca <- preprocess(gsca, species="Dm", initialIDs="FLYBASECG", keepMultipleMappings=TRUE, duplicateRemoverMethod="max", orderAbsValue=FALSE)
## do hypergeometric tests and GSEA
gsca <- analyze(gsca, para=list(pValueCutoff=0.05, pAdjustMethod ="BH", nPermutations=100, minGeneSetSize=200, exponent=1))
summarize(gsca)
## End(not run)
# ==================================================================
# NetWork Analysis
# Conducting the preparation steps
data(xn)
data(data4enrich)
# Conducting one sample t-test & compute the p-values
test.stats <- cellHTS2OutputStatTests(cellHTSobject=xn,annotationColumn="GeneID", alternative="two.sided",tests=c("T-test"))
library(BioNet)
pvalues <- BioNet::aggrPvals(test.stats, order=2, plot=FALSE)
nwa <- NWA(pvalues=pvalues, phenotypes=data4enrich)
nwa <- preprocess(nwa, species="Dm", initialIDs="FLYBASECG", keepMultipleMappings=TRUE, duplicateRemoverMethod="max")
nwa_inter <- interactome(nwa, species="Dm", reportDir="biogrid", genetic=FALSE)
# Case1: Conducting the network analysis step
nwa_result <- analyze(nwa_inter, fdr=0.0001, species="Dm")
# Case2: Conducting the network analysis step(with the interactionMatrix)
data(nwam)
nwam_inter <- interactome(nwam, species="Dm", reportDir="biogrid", genetic=FALSE)
nwam_result <- analyze(nwam_inter, fdr=0.0001, species="Dm")
}
