% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_analyze.R, R/nwa_analyze.R
\docType{methods}
\name{analyze,GSCA-method}
\alias{analyze,GSCA-method}
\alias{analyze,NWA-method}
\title{Gene Set Collection Analysis or NetWork Analysis}
\usage{
\S4method{analyze}{GSCA}(object, para = list(pValueCutoff = 0.05,
  pAdjustMethod = "BH", nPermutations = 1000, minGeneSetSize = 15, exponent =
  1), verbose = TRUE, doGSOA = FALSE, doGSEA = TRUE)

\S4method{analyze}{NWA}(object, fdr = 0.001, species, verbose = TRUE)
}
\arguments{
\item{object}{An object. When this function is implemented as the S4
method of class 'GSCA' or 'NWA', this argument is an object of class
'GSCA' or 'NWA'.}

\item{para}{A list of parameters for GSEA and hypergeometric tests.
This argument is only for class GSCA.}

\item{verbose}{A single logical value specifying to display detailed messages
(when verbose=TRUE) or not (when verbose=FALSE).}

\item{doGSOA}{A single logical value specifying to perform gene set
overrepresentation analysis(hypergeometric test) (when doGSOA=TRUE) or not (when doGSOA=FALSE).}

\item{doGSEA}{A single logical value specifying to perform gene set
enrichment analysis (when doGSEA=TRUE) or not (when doGSEA=FALSE).}

\item{fdr}{A single numeric value specifying the false discovery for the scoring of nodes
(see BioNet::scoreNodes and Dittrich et al., 2008 for details)}

\item{species}{A single character value specifying the species for which the data should be read.}

\item{para$pValueCutoff}{A single numeric value specifying the cutoff for adjusted p-values considered
significant.}

\item{para$pAdjustMethod}{A single character value specifying the p-value adjustment method to be used
(see 'p.adjust' for details).}

\item{para$nPermutations}{A single integer or numeric value specifying the number of permutations for
deriving p-values in GSEA.}

\item{para$minGeneSetSize}{A single integer or numeric value specifying the minimum number of elements
shared by a gene set and the input total genes. Gene sets with
fewer than this number are removed from both hypergeometric analysis and GSEA.}

\item{para$exponent}{A single integer or numeric value used in weighting phenotypes in GSEA.}
}
\value{
In the end, this function will return an updated object of
class GSCA or NWA. All the analyzed results could be found in slot \emph{result}.
}
\description{
This is a generic function.
}
\section{Methods (by class)}{
\itemize{
\item \code{GSCA}: The function will perform gene set collection analysis
on a GSCA object and update information about these results to slot
\emph{summary} of class GSCA.

\item \code{NWA}: The function will store the subnetwork module identified
by BioNet (if species is given, labels of nodes will also be mapped from
Entrez IDs to gene symbols), and update information about these results to
slot summary of class NWA.
}}

\examples{
# ====================================================
# Gene Set Collection Analysis Part
library(org.Hs.eg.db)
library(GO.db)
library(KEGGREST)
## load data for enrichment analyses
data(d7)
phenotype <- as.vector(d7$neg.lfc)
names(phenotype) <- d7$id

## select hits if you also want to do GSOA, otherwise ignore it
hits <-  names(phenotype[which(abs(phenotype) > 2)])

## set up a list of gene set collections
GO_BP <- GOGeneSets(species="Hs", ontologies=c("BP"))
PW_KEGG <- KeggGeneSets(species="Hs")
ListGSC <- list(GO_BP=GO_BP, PW_KEGG=PW_KEGG)

## create an object of class 'GSCA'
gsca <- new("GSCA", listOfGeneSetCollections = ListGSC, geneList = phenotype, hits = hits)

## do preprocessing
gsca1 <- preprocess(gsca, species="Hs", initialIDs="SYMBOL", keepMultipleMappings=TRUE,
                   duplicateRemoverMethod="max", orderAbsValue=FALSE)

## support parallel calculation using doParallel package
doParallel::registerDoParallel(cores=4)

## do hypergeometric tests and GSEA
gsca2 <- analyze(gsca1, para=list(pValueCutoff=0.05, pAdjustMethod ="BH",
                                nPermutations=100, minGeneSetSize=200, exponent=1),
                                doGSOA = TRUE, doGSEA = TRUE)

## summarize gsca2
summarize(gsca2)
head(gsca2@result$GSEA.results$GO_BP)
head(gsca2@result$HyperGeo.results$PW_KEGG)

# ===========================================================
# Network Analysis
library(org.Hs.eg.db)
library(GO.db)
## load data for network analyses
data(d7)
pvalues <- d7$neg.p.value
names(pvalues) <- d7$id

## input phenotypes if you want to color nodes by it
phenotypes <- as.vector(d7$neg.lfc)
names(phenotypes) <- d7$id

## create an object of class 'NWA' with phenotypes
nwa <- new("NWA", pvalues=pvalues, phenotypes=phenotypes)

## do preprocessing
nwa1 <- preprocess(nwa, species="Hs", initialIDs="SYMBOL", keepMultipleMappings=TRUE,
                  duplicateRemoverMethod="max")

## create an interactome for nwa by downloading for BioGRID database
nwa2 <- interactome(nwa1, species="Hs", reportDir="HTSanalyzerReport", genetic=FALSE)

## analyze
nwa3 <- analyze(nwa2, fdr=0.0001, species="Hs")

}
\references{
Aravind Subramanian, Pablo Tamayo, Vamsi K. Mootha, Sayan Mukherjee, Benjamin L. Ebert,
Michael A. Gillette, Amanda Paulovich, Scott L. Pomeroy, Todd R. Golub, Eric S. Lander, and Jill P. Mesirov
Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles
PNAS 2005 102 (43) 15545-15550; published ahead of print September 30, 2005, doi:10.1073/pnas.0506580102

Beisser D, Klau GW, Dandekar T, Muller T, Dittrich MT. BioNet: an R-Package
for the functional analysis of biological networks. Bioinformatics.
2010 Apr 15;26(8):1129-30.

Dittrich MT, Klau GW, Rosenwald A., Dandekar T and Muller T. Identifying functional modules
in protein-protein interaction networks: an integrated exact approach. Bioinformatics
2008 24(13):i223-i231.
}
