% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscaTS.R
\name{analyzeGscaTS}
\alias{analyzeGscaTS}
\title{Gene Set Collection Analysis for Time-series data.}
\usage{
analyzeGscaTS(gscaList, para = list(pValueCutoff = 0.05, pAdjustMethod = "BH",
  nPermutations = 1000, minGeneSetSize = 15, exponent = 1), verbose = TRUE,
  doGSOA = FALSE, doGSEA = TRUE)
}
\arguments{
\item{gscaList}{A named list of GSCA object}

\item{para}{A list of parameters for GSEA and hypergeometric tests. Details please see 'analyze'.}

\item{verbose}{a single logical value specifying to display detailed messages
(when verbose=TRUE) or not (when verbose=FALSE), default is TRUE.}

\item{doGSOA}{a single logical value specifying to perform gene set
overrepresentation analysis (when doGSOA=TRUE) or not (when doGSOA=FALSE),
default is FALSE.}

\item{doGSEA}{a single logical value specifying to perform gene set
enrichment analysis (when doGSEA=TRUE) or not (when doGSEA=FALSE),
default is TRUE.}
}
\value{
In the end, this function will return an updated list of GSCA object.
}
\description{
For each GSCA object in 'gscaList', this function will store the results from function
analyzeGeneSetCollections in slot result, and update information about
these results to slot summary of class GSCA.
}
\seealso{
\code{\link[HTSanalyzeR2]{analyze}}
}
