% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_preprocess.R
\name{annotationConvertor}
\alias{annotationConvertor}
\title{Convert between different types of gene identifiers}
\usage{
annotationConvertor(geneList, species = "Dm", initialIDs = "ENTREZID",
  finalIDs = "ENTREZID", keepMultipleMappings = TRUE, verbose = TRUE)
}
\description{
Convert between different types of gene identifiers
}

