% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_preprocess.R
\name{annotationConvertor}
\alias{annotationConvertor}
\title{Convert between different types of gene identifiers}
\usage{
annotationConvertor(geneList, species="Dm", initialIDs="FLYBASE",
finalIDs="Entrez.gene", keepMultipleMappings=TRUE, verbose=TRUE)
}
\arguments{
\item{geneList}{A named integer or numeric vector, or a matrix with
rows named by gene identifiers}

\item{species}{A single character value specifying the species for
which the data should be read.}

\item{initialIDs}{A single character value specifying the type of
initial identifiers for input geneList.}

\item{finalIDs}{A single character value specifying the type of
final identifiers for input geneList.}

\item{keepMultipleMappings}{A single logical value. If TRUE, the
function keeps the entries with multiple mappings (first mapping
is kept). If FALSE, the entries with multiple mappings will be discarded.}

\item{verbose}{A single logical value specifying to display detailed
messages (when verbose=TRUE) or not (when verbose=FALSE)}
}
\value{
The same data vector/matrix but with names/row names converted.
}
\description{
This function converts an initial data vector named by non-entrez
ids to the same vector but with entrez ids, and vice versa. Genes
for which no mapping were found will be removed. This function can
also take a matrix, with gene identifiers as row names.
}
\examples{
library(org.Dm.eg.db)
##example 1: convert a named vector
x<-runif(10)
names(x)<-names(as.list(org.Dm.egSYMBOL2EG))[1:10]
xEntrez<-annotationConvertor(geneList=x, species="Dm", initialIDs="SYMBOL",
finalIDs="ENTREZID")
##example 2: convert a data matrix with row names as gene ids
x<-cbind(runif(10),runif(10))
rownames(x)<-names(as.list(org.Dm.egSYMBOL2EG))[1:10]
xEntrez<-annotationConvertor(geneList=x, species="Dm", initialIDs="SYMBOL",
finalIDs="ENTREZID")

}
