% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscaTS.R
\name{appendGSTermsTS}
\alias{appendGSTermsTS}
\title{Append gene set terms to GSCA results for each GSCA object of Time-series data}
\usage{
appendGSTermsTS(gscaList, keggGSCs = NULL, goGSCs = NULL,
  msigdbGSCs = NULL)
}
\arguments{
\item{gscaList}{A named list of GSCA object.}

\item{keggGSCs}{a character vector of names of all KEGG gene set collections}

\item{goGSCs}{a character vector of names of all GO gene set collections}

\item{msigdbGSCs}{a character vector of names of all MSigDB gene set collections}
}
\description{
For each GSCA object in 'gscaList', this function
finds corresponding annotation terms for GO, KEGG and MSigDB gene sets and
inserts a column named "Gene.Set.Term" to each data frame in the GSCA results.
}
\seealso{
\code{\link[HTSanalyzeR2]{appendGSTerms}}
}
