% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Biogrid_HS_Interactome}
\alias{Biogrid_HS_Interactome}
\title{An object of class 'igraph'}
\format{An object of class \code{igraph} of length 10.}
\usage{
Biogrid_HS_Interactome
}
\description{
An object of class 'igraph' built from the interactome data set for
\emph{Homo Sapiens} downloaded from the \emph{BioGRID} database.
}
\keyword{datasets}
