% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d7}
\alias{d7}
\alias{d13}
\alias{d25}
\title{CRISPR data set}
\format{
An object of class \code{data.frame} with 8000 rows and 14 columns.
}
\usage{
d7
}
\description{
gRNA sequencing data after transducting the CRISPR system into a human cancer
cellline as time goes by. They are further pre-processed by MAGeCK. Details please
refer to \href{https://sourceforge.net/p/mageck/wiki/Home/}{MAGeCK}.
}
\examples{
data(d7)
data(d13)
data(d25)

}
\references{
Tzelepis K, Koike-Yusa H, De Braekeleer E, et al. A CRISPR Dropout Screen Identifies
Genetic Vulnerabilities and Therapeutic Targets in Acute Myeloid Leukemia. Cell Reports.
 2016;17(4):1193-1205. doi:10.1016/j.celrep.2016.09.079.
}
\keyword{datasets}
