% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_enrichmap.R
\docType{methods}
\name{extractEnrichMap,GSCA-method}
\alias{extractEnrichMap,GSCA-method}
\title{Extract the enrichment map result from GSCA object}
\usage{
\S4method{extractEnrichMap}{GSCA}(object, resultName = "GSEA.results", gscs,
  ntop = NULL, allSig = TRUE, gsNameType = "id", specificGeneset = NULL)
}
\description{
This is a generic function.
}
