% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_enrichmap.R
\name{fetchGSCASeriesValues}
\alias{fetchGSCASeriesValues}
\title{Extract and combine enrichment map information of given GSCA objects.}
\usage{
fetchGSCASeriesValues(gscaObjs, resultName = "GSEA.results", gscs,
  ntop = NULL, allSig = TRUE, gsNameType = "id", specificGeneset = NULL,
  cutoff = NULL)
}
\description{
This method is internally used for force-graph drawing. The enrichment map data of all the GSCA objects
are extacted and combined into corresponding dataframes. The attributes that are changes with the time
are also returned.
}
