% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3plot.R
\name{forceGraph}
\alias{forceGraph}
\title{Create a force-directed graph widget}
\usage{
forceGraph(
  nodes,
  links,
  nMappings,
  lMappings,
  options,
  width = NULL,
  height = NULL,
  seriesData = NULL
)
}
\arguments{
\item{nodes}{the node information of force-directed graph}

\item{links}{the link information of force-directed graph}

\item{nMappings}{indicates which attributes are used for node rendering. The available render
options are "id", "size", "color", "scheme", "label", "label_id" and "label_term".}

\item{lMappings}{indicates which attributes are used for link rendering. The available render
options are "source", "target" and "weight".}

\item{options}{A list of internal meta configurations provided for JavaScript library. Users
are highly recommended to modify the detailed rendering options in the shiny report by
\code{\link[HTSanalyzeR2]{report}}.}

\item{width}{the width of the widget}

\item{height}{the height of the widget}

\item{seriesData}{A list of time points in time series analysis.}
}
\description{
This function creates a force-directed graph using \pkg{htmlwidgets}, which can be rendered
in R Markdown documents and shiny applications.
The rendering is conducted using a variant of the Sigma which is a JavaScript library dedicated
to graph drawing. The force-graph layout algorithm is ForceAtlas2
[http://www.plosone.org/article/info%3Adoi%2F10.1371%2Fjournal.pone.0098679] and implemented by
Mathieu Jacomy [https://github.com/jacomyma], Guillaume Plique [https://github.com/Yomguithereal]
and Sébastien Heymann. [https://github.com/sheymann].
}
