% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_load_genesets.R
\name{GOGeneSets}
\alias{GOGeneSets}
\title{Create a list of gene sets based on GO terms}
\usage{
GOGeneSets(species = "Dm", ontologies = c("MF"))
}
\arguments{
\item{species}{A single character value specifying a choice of species,
such as "Dm" ("Drosophila_melanogaster"), "Hs" ("Homo_sapiens"),
"Rn" ("Rattus_norvegicus") or "Mm" ("Mus_musculus").}

\item{ontologies}{A single character value or a character vector
specifying an ontology or multiple ontologies. The current version
provides the following choices: "BP", "CC" and "MF"}
}
\value{
A list of gene sets, with names as GO IDs. Each gene set is
a character vector of Entrez identifiers.
}
\description{
This function creates a list of gene sets based on GO terms.
It is species-specific, and returns a list of gene sets, each
of which is a character vector of Entrez identifiers.
}
\details{
This function relies on the following packages:
AnnotationDbi, GO.db and the species db, such as org.Dm.eg.db.
}
\examples{
library(GO.db)
library(org.Dm.eg.db)
DM_GO_CC<-GOGeneSets(species="Dm",ontologies=c("CC"))

}
\seealso{
KeggGeneSets, MSigDBGeneSets
}

