% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsca}
\alias{gsca}
\title{An object of class 'GSCA'}
\format{An object of class \code{GSCA} of length 1.}
\usage{
gsca
}
\description{
An object of class 'GSCA' generated by \emph{HTSanalyzeR2}.
}
\keyword{datasets}
