% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_class.R
\name{GSCA}
\alias{GSCA}
\title{An S4 class for Gene Set Collection Analyses on high-throughput data}
\usage{
GSCA(listOfGeneSetCollections, geneList, hits = character())
}
\description{
This S4 class includes a series of methods to do gene set enrichment analysis
and hypergeometric test for high-throughput data.
}
\section{Slots}{

\describe{
\item{\code{listOfGeneSetCollections}}{A list of gene set collections (a 'gene
set collection' is a list of gene sets).}

\item{\code{geneList}}{A numeric or integer vector of phenotypes named by gene
identifiers.}

\item{\code{hits}}{A character vector of the gene identifiers (used as hits in
the hypergeometric tests).It's needed if you want to do GSOA.}

\item{\code{para}}{A list of parameters for hypergeometric test and GSEA. These
parameters are pValueCutoff, pAdjustMethod, nPermutations, minGeneSetSize
and exponent.}

\item{\code{result}}{A list of results.}

\item{\code{summary}}{A list of summary information for listOfGeneSetCollections,
geneList, hits, para, and result.}

\item{\code{preprocessed}}{A single logical value specifying whether or not the
input data has been preprocessed.}
}}

\seealso{
\code{\link[HTSanalyzeR2]{preprocess}},
\code{\link[HTSanalyzeR2]{analyze}},
\code{\link[HTSanalyzeR2]{appendGSTerms}},
\code{\link[HTSanalyzeR2]{summarize}},
\code{\link[HTSanalyzeR2]{report}}
}
