% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gscaTS.R
\name{gscaTS}
\alias{gscaTS}
\title{Produce a list of GSCA object for Gene Set Collection Analyses on high-throughput screens}
\usage{
gscaTS(object, listOfGeneSetCollections)
}
\arguments{
\item{object}{A TSImport object.}

\item{listOfGeneSetCollections}{A list of gene set collections (a 'gene
set collection' is a list of gene sets).}
}
\value{
In the end, this function will return a named list of new GSCA object.
}
\description{
This function will produce a list of GSCA object for time-series data to do Gene Set Enrichment Analyses
and hypergeometric analysis
}
\seealso{
\code{\link[HTSanalyzeR2]{GSCA}}
}
