% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwa_preprocess.R
\docType{methods}
\name{interactome,NWA-method}
\alias{interactome,NWA-method}
\title{Create an interactome from BioGRID data sets}
\usage{
\S4method{interactome}{NWA}(object, interactionMatrix = NULL, species,
  link = "http://thebiogrid.org/downloads/archives/Release\%20Archive/BIOGRID-3.4.138/BIOGRID-ORGANISM-3.4.138.tab2.zip",
  reportDir = "HTSanalyzerReport", genetic = FALSE, force = FALSE,
  verbose = TRUE)
}
\arguments{
\item{object}{an object. When this function is implemented as the S4
method of class NWA, this argument is an object of class 'NWA'}

\item{interactionMatrix}{an interaction matrix including columns
'InteractionType', 'InteractorA' and 'InteractorB'. If this matrix
is available, the interactome can be directly built based on it.}

\item{species}{a single character value specifying the species for
which the data should be read.}

\item{link}{the link (url) where the data should be downloaded (in
tab2 format). The default link is version 3.4.138}

\item{reportDir}{a single character value specifying the directory
to store reports. The BioGRID data set will be downloaded and stored
in a subdirectory called 'Data' in 'reportDir'.}

\item{genetic}{a single logical value. If TRUE, genetic interactions
will be kept; otherwise, they will be removed from the data set.}

\item{force}{force to download the data set.}

\item{verbose}{a single logical value indicating to display detailed
messages (when verbose=TRUE) or not (when verbose=FALSE)}
}
\description{
This is a generic function.
When implemented as the S4 method of class NWA, this function creates an
interactome before conducting network analysis.
}

