% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwaTS.R
\name{interactomeNwaTS}
\alias{interactomeNwaTS}
\title{Create an interactome from BioGRID data sets}
\usage{
interactomeNwaTS(nwaList, interactionMatrix = NULL, species,
  link = "http://thebiogrid.org/downloads/archives/Release\%20Archive/BIOGRID-3.4.138/BIOGRID-ORGANISM-3.4.138.tab2.zip",
  reportDir = "HTSanalyzerReport", genetic = FALSE, force = FALSE,
  verbose = TRUE)
}
\arguments{
\item{nwaList}{A named list of NWA object generated by preprocessNwaTS.}

\item{interactionMatrix}{an interaction matrix including columns
'InteractionType', 'InteractorA' and 'InteractorB'. If this matrix
is available, the interactome can be directly built based on it.}

\item{species}{a single character value specifying the species for
which the data should be read.}

\item{link}{the link (url) where the data should be downloaded (in
tab2 format). The default link is version 3.4.138}

\item{reportDir}{a single character value specifying the directory
to store reports. The BioGRID data set will be downloaded and stored
in a subdirectory called 'Data' in 'reportDir'.}

\item{genetic}{a single logical value. If TRUE, genetic interactions
will be kept; otherwise, they will be removed from the data set.}

\item{force}{force to download the data set.}

\item{verbose}{a single logical value indicating to display detailed
messages (when verbose=TRUE) or not (when verbose=FALSE)}
}
\value{
In the end, this function will return an updated list of NWA object.
}
\description{
This function creates an interactome before conducting network analysis for Time-series data
}
\seealso{
\code{\link[HTSanalyzeR2]{interactome}}
}
