% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_load_genesets.R
\name{KeggGeneSets}
\alias{KeggGeneSets}
\title{Create a list of gene sets based on KEGG pathways terms}
\usage{
KeggGeneSets(species = "Dm")
}
\arguments{
\item{species}{A single character value specifying a choice of species,
such as "Dm" ("Drosophila_melanogaster"), "Hs" ("Homo_sapiens"),
"Rn" ("Rattus_norvegicus") or "Mm" ("Mus_musculus").}
}
\value{
A list of gene sets, with names as KEGG pathway IDs. Each gene
set is a character vector of Entrez gene identifiers.
}
\description{
This function creates a list of gene sets based on KEGG pathways terms.
It is species-specific, and returns a list of gene sets, each of which
is a character vector of Entrez gene identifiers.
}
\details{
This function needs Internet connection and relies on the
following packages: KEGGREST
}
\examples{
library(KEGGREST)
DM_KEGG<-KeggGeneSets(species = "Dm")

}
\seealso{
\code{\link[HTSanalyzeR2]{GOGeneSets}}, \code{\link[HTSanalyzeR2]{MSigDBGeneSets}}
}

