% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_load_genesets.R
\name{MSigDBGeneSets}
\alias{MSigDBGeneSets}
\title{Create a list of gene sets based on MSigDB collection terms}
\usage{
MSigDBGeneSets(collection = "h")
}
\arguments{
\item{collection}{A single character value specifying a choice of collection.}
}
\value{
Return A list of gene sets.
}
\description{
This function creates gene set collections based on MSigDB.
It is collection-specific, and returns a GeneSetCollection objects with
the elements of the gene sets represented by Entrez Gene IDs.
}
\examples{
H_MSig <- MSigDBGeneSets(collection = "h")

}
\seealso{
GOGeneSets, KeggGeneSets
}

