% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_load_genesets.R
\name{MSigDBGeneSets}
\alias{MSigDBGeneSets}
\title{Create a list of gene sets for specific species from MSigDB database}
\usage{
MSigDBGeneSets(collection = "c2", species = "Hs")
}
\arguments{
\item{collection}{A single character value specifying a choice of collection.
Valid values include 'h'(hallmark gene sets), 'c1' (positional gene sets),
'c2' (curated gene sets), 'c3' (motif gene sets), 'c4' (computational gene sets),
'c5' (GO gene sets), 'c6' (oncogenic signatures), 'c7' (immunologic signatures).
More details please refer to
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB}.}

\item{species}{A single character value specifying the species of the gene sets of MSigDB.
Now we only support 'Hs' (Homo_Sapiens) for all 8 gene set collections
and 'Mm' (Mus_musculus) for 'c2', 'c6' and 'c7' gene set collections.}
}
\value{
Return a list of gene sets of specific collection in
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB} of version 6.1.
}
\description{
This function creates gene set collections based on MSigDB database version 6.1. Currently
our package only supports all 8 collections for Homo Sapiens as well as
'c2', 'c6' and 'c7' for Mus musculus. It returns a list of gene sets collections
with the elements of the gene sets represented by Entrez Gene IDs.
}
\examples{
C2_MSig <- MSigDBGeneSets(collection = "c2", species = "Hs")
}
\seealso{
\code{\link[HTSanalyzeR2]{GOGeneSets}}, \code{\link[HTSanalyzeR2]{KeggGeneSets}}
}
