% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_load_genesets.R
\name{MSigDBGeneSets}
\alias{MSigDBGeneSets}
\title{Create a list of gene sets based on MSigDB collection terms}
\usage{
MSigDBGeneSets(collection = "c2")
}
\arguments{
\item{collection}{A single character value specifying a choice of collection.
It should be either 'c2'(curated gene sets) or 'c5'(GO gene sets). More details please refer to
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB}.}
}
\value{
Return a list of gene sets of specific collection in
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB}.
}
\description{
This function creates gene set collections based on MSigDB. Currently only
commonly used 'c2' and 'c5' gene sets could be gotten by this function.
It is collection-specific, and returns a list of gene sets collection with
the elements of the gene sets represented by Entrez Gene IDs.
}
\examples{
C2_MSig <- MSigDBGeneSets(collection = "c2")

}
\seealso{
\code{\link[HTSanalyzeR2]{GOGeneSets}}, \code{\link[HTSanalyzeR2]{KeggGeneSets}}
}
