% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_load_genesets.R
\name{MSigDBGeneSets}
\alias{MSigDBGeneSets}
\title{Create a list of gene sets based on MSigDB collection terms}
\usage{
MSigDBGeneSets(collection = "h")
}
\arguments{
\item{collection}{A single character value specifying a choice of collection.
It should be one of the following character strings: 'h'(hallmark gene sets),
'c1'(positional gene sets), 'c2'(curated gene sets), 'c3'(motif gene sets),
'c4'(computational gene sets), 'c5'(GO gene sets), 'c6'(oncogenic signatures),
'c7'(immunologic signatures). More details please refer to
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB}.}
}
\value{
Return a list of gene sets of specific collection in
\href{http://software.broadinstitute.org/gsea/msigdb}{MSigDB}.
}
\description{
This function creates gene set collections based on MSigDB.
It is collection-specific, and returns a list of gene sets collection with
the elements of the gene sets represented by Entrez Gene IDs.
}
\examples{
\dontrun{
H_MSig <- MSigDBGeneSets(collection = "h")
}

}
\seealso{
\code{\link[HTSanalyzeR2]{GOGeneSets}}, \code{\link[HTSanalyzeR2]{KeggGeneSets}}
}
