% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwa_class.R
\name{NWA}
\alias{NWA}
\title{An S4 class for NetWork Analysis on high-throughput screens}
\usage{
NWA(pvalues, phenotypes = as.numeric(NA), interactome = NA)
}
\description{
This class includes a series of methods to do network analysis for
high-throughput screens.
}
\section{Slots}{

\describe{
\item{\code{pvalues}}{a numeric vector of p-values.}

\item{\code{phenotypes}}{a numeric or integer vector of phenotypes.}

\item{\code{interactome}}{an object of class igraph.}

\item{\code{fdr}}{one parameter for BioNet to score nodes in the interactome.}

\item{\code{result}}{a list consisting of subnetwork module identified by BioNet
and a vector of labels for nodes of the subnetwork module.}

\item{\code{summary}}{a list of summary information for p-values, phenotypes,
interactome and result.}

\item{\code{preprocessed}}{a logical value specifying whether or not input data
has been preprocessed.}
}}
\seealso{
preprocess analyze summarize interactome report
}

