% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwaTS}
\alias{nwaTS}
\title{A list of 'NWA' object}
\format{An object of class \code{list} of length 3.}
\usage{
nwaTS
}
\description{
A list of 'NWA' object generated by time-series analysis using \emph{HTSanalyzeR2}.
}
\examples{
data(nwaTS)
}
\keyword{datasets}
