% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_view.R
\docType{methods}
\name{plotGSEA,GSCA-method}
\alias{plotGSEA,GSCA-method}
\title{Plot and save figures of GSEA results for top significant gene sets}
\usage{
\S4method{plotGSEA}{GSCA}(object, gscs, ntop = NULL, allSig = FALSE,
  filepath = ".", output = "pdf", ...)
}
\arguments{
\item{object}{an object. When this function is implemented as the S4 method of class GSCA,
this argument is an object of class GSCA.}

\item{gscs}{a character vector specifying the names of gene set collections whose top
significant gene sets will be plotted}

\item{ntop}{a single integer or numeric value specifying how many gene sets of top
significance will be plotted.}

\item{allSig}{a single logical value. If 'TRUE', all significant gene sets (GSEA adjusted
p-value < 'pValueCutoff' of slot 'para') will be plotted; otherwise, only top 'ntop' gene
 sets will be plotted.}

\item{filepath}{a single character value specifying where to store GSEA figures.}

\item{output}{a single character value specifying the format of output image: "pdf" or "png"}

\item{...}{other arguments used by the function png or pdf such as 'width' and 'height'}
}
\description{
This is a generic function.When implemented as the S4 method for objects of class GSCA,
this function plots figures of the positions of genes of the gene set in the ranked gene
list and the location of the enrichment score for top significant gene sets.
}
\examples{
## Not run:
library(org.Dm.eg.db)
library(GO.db)
## load data for enrichment analyses
data(data4enrich)
## select hits
hits <- names(data4enrich)[abs(data4enrich) > 2]
## set up a list of gene set collections
GO_MF <- GOGeneSets(species="Dm", ontologies=c("MF"))
PW_KEGG <- KeggGeneSets(species="Dm")
H_MSig <- MSigDBGeneSets(collection = "h")
ListGSC <- list(GO_MF=GO_MF, H_MSig = H_MSig, PW_KEGG=PW_KEGG)
## create an object of class 'GSCA'
gsca <- GSCA(listOfGeneSetCollections = ListGSC, geneList = data4enrich, hits = hits)
## print gsca
gsca
## do preprocessing
gsca <- preprocess(gsca, species="Dm", initialIDs="FLYBASECG", keepMultipleMappings=TRUE, duplicateRemoverMethod="max", orderAbsValue=FALSE)
## do hypergeometric tests and GSEA
gsca <- analyze(gsca, para=list(pValueCutoff=0.05, pAdjustMethod ="BH", nPermutations=100, minGeneSetSize=200, exponent=1))
summarize(gsca)
## plot  significant gene sets in GO_MF and PW_KEGG
plotGSEA(gsca, gscs=c("GO_MF","PW_KEGG"), ntop=1, filepath=".")
}
