% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_view.R
\docType{methods}
\name{plotGSEA,GSCA-method}
\alias{plotGSEA,GSCA-method}
\alias{plotGSEA}
\title{Plot and save figures of GSEA results for top significant gene sets}
\usage{
\S4method{plotGSEA}{GSCA}(object, gscs, ntop = NULL, allSig = FALSE,
  filepath = ".", output = "pdf", ...)
}
\arguments{
\item{object}{A GSCA object.}

\item{gscs}{A character vector specifying the names of gene set collections whose top
significant gene sets will be plotted.}

\item{ntop}{A single integer or numeric value specifying how many gene sets of top
significance will be plotted.}

\item{allSig}{A single logical value. If 'TRUE', all significant gene sets (GSEA adjusted
p-value < 'pValueCutoff' of slot 'para') will be plotted; otherwise, only top 'ntop' gene
 sets will be plotted.}

\item{filepath}{A single character value specifying where to store GSEA figures.}

\item{output}{A single character value specifying the format of output image: "pdf" or "png".}

\item{...}{Other arguments used by the function png or pdf such as 'width' and 'height'}
}
\value{
In the end, this function would plot GSEA figure and store them into the specified path.
}
\description{
This is a generic function.When implemented as the S4 method for objects of class GSCA,
this function plots figures of the positions of genes of the gene set in the ranked gene
list and the location of the enrichment score for top significant gene sets.
}
\examples{
## load a GSCA object(see the examples of analyze GSCA for details)
data(d7_gsca)

## summarize d7_gsca
summarize(d7_gsca)

## plot  significant gene sets in GO_MF and PW_KEGG
\dontrun{
plotGSEA(d7_gsca, gscs=c("GO_MF","PW_KEGG"), ntop=3, filepath=".")
}
}
