% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_report.R, R/nwa_report.R
\docType{methods}
\name{report,GSCA-method}
\alias{report,GSCA-method}
\alias{report}
\alias{report,NWA-method}
\title{Write HTML reports for enrichment or network analyses}
\usage{
\S4method{report}{GSCA}(object, specificGeneset = NULL,
  reportDir = "GSCAReport")

\S4method{report}{NWA}(object, reportDir = "NWAReport")
}
\arguments{
\item{object}{An object. When this function is implemented as the S4
method of class 'GSCA' or 'NWA', this argument is an object of class
'GSCA' or 'NWA'.}

\item{specificGeneset}{A named list of specific gene sets. See \code{\link[HTSanalyzeR2]{viewEnrichMap,GSCA-method}}
for details.}

\item{reportDir}{A single character value specifying the directory to store reports. For default
the enrichment analysis reports will be stored in the directory called "GSCAReport".}
}
\value{
in the end, this function would generate a shiny report.

in the end, this function would generate a shiny report.
}
\description{
This is a generic function.
}
\details{
When implemented as the method of class GSCA or NWA, this function produces reports for
either the Gene Set Collection Analysis or the NetWork Analysis.


This will generate a shiny report including all the GSCA or NWA results.
For GSCA object, users can download the table of GSOA and/or GSEA result in different format
such as 'csv' and 'pdf'. The enrichment map could be modified according to the user's preferences
such as layout, node, label, and etc. Details please see the vignette of our package.

For NWA object, the identified subnetwork could be modified according to the user's preferences
in many ways such as layout, color, and etc. Details please see the vignette of our package.
}
\examples{
# =======================================================
\dontrun{
# GSCA class
## load a GSCA object(see the examples of analyze GSCA for details)
data(d7_gsca)

## Example1: report d7_gsca
report(d7_gsca)

## Example2: report d7_gsca containing enrichment map with specificGeneset
tmp <- getTopGeneSets(d7_gsca, resultName = "GSEA.results", gscs=c("GO_MF"),
                      ntop = 20000, allSig = FALSE)
## In that case, we can define specificGeneset as below:
GO_MF_geneset <- tmp$GO_MF[c(4,2,6,9,12)]
## the name of specificGenesets also needs to match with the names of tmp
specificGeneset <- list("GO_MF"=GO_MF_geneset)
report(d7_gsca, specificGeneset=specificGeneset)
}
\dontrun{
# =================================================================
# NWA class
## load a NWA object(see the examples of analyze NWA for details)
data(d7_nwa)

## report d7_nwa
report(d7_nwa)
}
}
