% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_report.R
\name{reportAll}
\alias{reportAll}
\title{Write HTML reports for both enrichment and network analyses}
\usage{
reportAll(gsca = NULL, nwa = NULL, TSOrder = NULL,
  specificGeneset = NULL, reportDir = "AnalysisReport")
}
\arguments{
\item{gsca}{a GSCA object or a list of GSCA objects}

\item{nwa}{a NWA object or a list of NWA objects}

\item{TSOrder}{A character specifying the visulization order of 'Time Series' data in shiny report. Only works when
reporting for 'Time Series' data, default is the ID order in 'expInfor'}

\item{reportDir}{a single character value specifying the directory to store reports. For default both the
enrichment analysis and network analysis reports will be stored in the directory called "AnalysisReport"}
}
\description{
This function can create HTML reports for both gene sets enrichment analysis and network analysis using
shiny.
}
