% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_summarize.R, R/nwa_summarize.R
\docType{methods}
\name{summarize,GSCA-method}
\alias{summarize,GSCA-method}
\alias{summarize,NWA-method}
\title{Print summary information for an object of class GSCA or NWA}
\usage{
\S4method{summarize}{GSCA}(object, what = "ALL")

\S4method{summarize}{NWA}(object, what = "ALL")
}
\arguments{
\item{object}{an object. When this function is implemented as the S4 method
of class GSCA or NWA, this argument is an object of class GSCA or NWA.}

\item{what}{a single character value or a character vector of key words
specifying what to print (see details below).}
}
\description{
This is a generic function.
When implemented as the S4 method for objects of class GSCA or NWA, this
function prints a summary of information about the slots of these classes.
}
\section{Methods (by class)}{
\itemize{
\item \code{GSCA}: For an object of class GSCA, the key words are 'GSC'
(the slot 'listOfGeneSetCollections'), 'GeneList' (the slot 'geneList'),
'Hits' (the slot 'hits'), 'Para' (the slot 'para'), 'Result' (the slot
'result') and 'ALL' (all slots).

\item \code{NWA}: For an object of class NWA, the key words include 'Pval'
(the slot 'pvalue'), 'Phenotype' (the slot 'phenotype'), 'Interactome' (the
slot 'interactome'), 'Para' (the slot 'fdr'), 'Result' (the slot 'result')
and 'ALL' (all slots).
}}

