% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_enrichmap.R
\docType{methods}
\name{viewEnrichMap,GSCA-method}
\alias{viewEnrichMap,GSCA-method}
\title{Plot a figure of the enrichment map for GSEA or Hypergeometric tests}
\usage{
\S4method{viewEnrichMap}{GSCA}(object, resultName = "GSEA.results", gscs,
  ntop = NULL, allSig = TRUE, gsNameType = "id", specificGeneset = NULL,
  options = list(), seriesObjs = NULL)
}
\arguments{
\item{object}{A GSCA object.}

\item{resultName}{A single character value specifying draw an enrichment map based on
which result. Could only be 'HyperGeo.results' or 'GSEA.results'.}

\item{gscs}{A character vector specifying the names of gene set collections
of which the top significant gene sets will be plotted.}

\item{ntop}{A single integer or numeric value specifying how many gene
sets of top significance will be plotted.}

\item{allSig}{A single logical value. If 'TRUE', all significant gene sets
(GSEA adjusted p-value < 'pValueCutoff' of slot 'para') will be used; otherwise,
only top 'ntop' gene sets will be used.}

\item{gsNameType}{A single character value specifying the type of the gene set names that
will be displayed as the names of nodes in the enrichment map. The type of the gene
set names should be one of the following: "id", "term" or "none".}

\item{specificGeneset}{A named list of specific gene sets. Specifically, this term needs to be
a subset of all analyzed gene sets which can be roughly gotten by
\strong{getTopGeneSets(object, resultName, gscs, ntop = 20000, allSig = FALSE)} with warning.
The order of the list must match the order of results gotten by aboved function \strong{getTopGeneSets}.}
}
\value{
An object of igraph with all attributes about the enrichement map.
}
\description{
This is a generic function. When implemented as the S4 method for objects
of class GSCA, this function will plot an enrichment map for GSEA or
Hypergeometric test results.
}
\details{
The idea of this function is similar to the PLoS one paper by Merico et al.

An enrichment map is a network to help better visualize and interpret the GSEA or
Hypergeometric test results. In an enrichment map, the nodes represent gene sets
and the edges denote the Jaccard similarity coefficient between two gene sets.
Node colors are scaled according to the adjusted p-values (the darker the more significant).
For GSEA, nodes are colored by the sign of the enrichment scores (red:+, blue: -).
The size of nodes illustrates the size of gene sets, while the width of edges denotes the
Jaccard coefficient.

A useful application of this function is that user can define specificGeneset to plot an
enrichment map with their interested gene sets instead of the top significant ones or all the significant
ones. Especially when the number of the significant gene sets are huge, which would make the enrichment
map a mess as well as of no information. To this end, user can set the parameter \strong{specificGeneset}.
}
\examples{
library(org.Hs.eg.db)
library(GO.db)
library(KEGGREST)
## load data for enrichment analyses
data(d7)
phenotype <- as.vector(d7$neg.lfc)
names(phenotype) <- d7$id

## select hits if you also want to do GSOA, otherwise ignore it
hits <-  names(phenotype[which(abs(phenotype) > 2)])

## set up a list of gene set collections
GO_MF <- GOGeneSets(species="Hs", ontologies=c("MF"))
PW_KEGG <- KeggGeneSets(species="Hs")
ListGSC <- list(GO_MF=GO_MF, PW_KEGG=PW_KEGG)

## create an object of class 'GSCA'
gsca <- new("GSCA", listOfGeneSetCollections = ListGSC, geneList = phenotype, hits = hits)

## do preprocessing
gsca <- preprocess(gsca, species="Hs", initialIDs="SYMBOL", keepMultipleMappings=TRUE,
                   duplicateRemoverMethod="max", orderAbsValue=FALSE)

## support parallel calculation using doParallel package
doParallel::registerDoParallel(cores=4)

## do hypergeometric tests and GSEA
gsca <- analyze(gsca, para=list(pValueCutoff=0.05, pAdjustMethod ="BH",
                                nPermutations=100, minGeneSetSize=200, exponent=1),
                                doGSOA = TRUE, doGSEA = TRUE)

## append gene set terms to results
gsca <- appendGSTerms(gsca, goGSCs=c("GO_MF"), keggGSCs=NULL, msigdbGSCs=NULL)


#' ## view an enrichment map for top 7 significant 'GO_MF' gene sets of GSEA results
library(igraph)
viewEnrichMap(gsca, resultName = "GSEA.results", gscs=c("GO_MF"),
              allSig = F, ntop = 7, gsNameType = "term")

## view an enrichment map for top 7 significant 'GO_MF' and 'PW_KEGG' gene sets of GSEA results
library(igraph)
viewEnrichMap(gsca, resultName = "GSEA.results", gscs=c("GO_MF", "PW_KEGG"),
              allSig = F, ntop = 7, gsNameType = "term")

## view an enrichment map with specificGenesets in 'GO_MF' gene sets of GSEA results
library(igraph)
## As told previously, specificGeneset needs to be a subset of all analyzed gene sets
## which can be roughly gotten by getTopGeneSets(object, resultName, gscs, ntop = 20000, allSig = FALSE)
tmp <- getTopGeneSets(gsca, resultName = "GSEA.results", gscs=c("GO_MF"), ntop = 20000, allSig = FALSE)
In that case, we can define specificGeneset as below:
GO_MF_geneset <- tmp$GO_MF[c(4,2,6,9,12)]
## the name of specificGenesets also needs to match with the names of tmp
specificGeneset <- list("GO_MF"=GO_MF_geneset)
viewEnrichMap(gsca, resultName = "GSEA.results", gscs=c("GO_MF"), allSig = F, gsNameType = "term",
              ntop = NULL, specificGeneset = specificGeneset)


## view an enrichment map with specificGenesets in 'GO_MF' and 'PW_KEGG' gene sets of GSEA results
tmp <- getTopGeneSets(gsca, resultName = "GSEA.results", gscs=c("GO_MF", "PW_KEGG"),
                      ntop = 20000, allSig = FALSE)
GO_MF_geneset <- tmp$GO_MF[c(6,3,5,9,12)]
PW_KEGG_geneset <- tmp$PW_KEGG[c(7,2,5,1,9)]
specificGeneset <- list("GO_MF"=GO_MF_geneset, "PW_KEGG"=PW_KEGG_geneset)
viewEnrichMap(gsca, resultName = "GSEA.results", gscs=c("GO_MF", "PW_KEGG"), allSig = F,
              gsNameType = "term", ntop = NULL, specificGeneset = specificGeneset)


}
