% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_view.R
\docType{methods}
\name{viewGSEA,GSCA-method}
\alias{viewGSEA,GSCA-method}
\alias{viewGSEA}
\title{Plot a figure of GSEA results for one gene set}
\usage{
\S4method{viewGSEA}{GSCA}(object, gscName, gsName)
}
\arguments{
\item{object}{A GSCA object.}

\item{gscName}{A single character value specifying the name of the gene set collection where
the gene set is.}

\item{gsName}{A single character value specifying the name of the gene set to be plotted.}
}
\value{
In the end, this function would draw a GSEA figure for specified gene set.
}
\description{
This is a generic function. When implemented as the S4 method for objects of class GSCA,
 this function plots a figure
of the positions of the gene sets in the ranked gene list and the location of the enrichment score.
}
\details{
We suggest to print the names of top significant gene sets using the
function \code{\link[HTSanalyzeR2]{getTopGeneSets}} before plotting the GSEA results.
}
\examples{
## load a GSCA object(see the examples of analyze GSCA for details)
data(d7_gsca)

## summarize gsca
summarize(d7_gsca)

## print top significant gene sets in GO_MF
topGS_GO_MF <- getTopGeneSets(d7_gsca, "GSEA.results", gscs = "GO_MF", allSig=TRUE)

## view GSEA results for one gene set
viewGSEA(d7_gsca, "GO_MF", topGS_GO_MF[["GO_MF"]][1])
}
