% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsca_view.R
\docType{methods}
\name{viewGSEA,GSCA-method}
\alias{viewGSEA,GSCA-method}
\alias{viewGSEA}
\title{Plot a figure of GSEA results for one gene set}
\usage{
\S4method{viewGSEA}{GSCA}(object, gscName, gsName)
}
\arguments{
\item{object}{A GSCA object.}

\item{gscName}{A single character value specifying the name of the gene set collection where
the gene set is.}

\item{gsName}{A single character value specifying the name of the gene set to be plotted.}
}
\description{
This is a generic function. When implemented as the S4 method for objects of class GSCA,
 this function plots a figure
of the positions of the gene sets in the ranked gene list and the location of the enrichment score.
}
\details{
We suggest to print the names of top significant gene sets using the
function \code{\link[HTSanalyzeR2]{getTopGeneSets}} before plotting the GSEA results.
}
\examples{
\dontrun{
## load a GSCA object(see the examples of analyze GSCA for details)
data(gsca)

## summarize gsca
summarize(gsca)

## print top significant gene sets in GO_BP
topGS_GO_BP <- getTopGeneSets(gsca, "GSEA.results", gscs = "GO_BP", allSig=TRUE)

## view GSEA results for one gene set
viewGSEA(gsca, "GO_BP", topGS_GO_BP[["GO_BP"]][1])
}
}
